keystore: fix uid buffer length
The IOCTL_GET_DEVICE_UID ioctl expects a 32-byte buffer,
so correct the length of uid[] and update the HMAC
generation to use only the first 16 bytes of it.

The previous 16-byte buffer worked OK before because the
last 16 bytes of the UID are all zero, and the variables
after uid[] in storage were expected to be zero anyway.

Signed-off-by: Matt Madison <matt@madison.systems>